refactor: move tx-details tests
