+++ ui/pages/bridge/transaction-details/transaction-details.test.tsx
@@ -11,7 +11,6 @@ import { mockNetworkState } from '../../../../test/stub/networks';
 import { CHAIN_IDS } from '../../../../shared/constants/network';
 import configureStore from '../../../store/store';
 import { TransactionGroup } from '../../../hooks/useTransactionDisplayData';
-import { getIsDelayed } from '../utils/tx-details';
 import CrossChainSwapTxDetails from './transaction-details';
 
 jest.mock('react-router-dom', () => ({
@@ -68,60 +67,6 @@ const getMockStore = (
 };
 
 describe('transaction-details', () => {
-  describe('getIsDelayed', () => {
-    it('returns false when status is not PENDING', () => {
-      const result = getIsDelayed(StatusTypes.COMPLETE, {
-        startTime: Date.now(),
-        estimatedProcessingTimeInSeconds: 60,
-      } as BridgeHistoryItem);
-      expect(result).toBe(false);
-    });
-
-    it('returns false when bridgeHistoryItem is undefined', () => {
-      const result = getIsDelayed(StatusTypes.PENDING, undefined);
-      expect(result).toBe(false);
-    });
-
-    it('returns false when startTime is undefined', () => {
-      const result = getIsDelayed(StatusTypes.PENDING, {
-        startTime: undefined,
-        estimatedProcessingTimeInSeconds: 60,
-      } as BridgeHistoryItem);
-      expect(result).toBe(false);
-    });
-
-    it('returns false when current time is less than estimated completion time', () => {
-      const result = getIsDelayed(StatusTypes.PENDING, {
-        startTime: Date.now() - 1000,
-        estimatedProcessingTimeInSeconds: 60,
-      } as BridgeHistoryItem);
-
-      expect(result).toBe(false);
-    });
-
-    it('returns true when current time exceeds estimated completion time by 10 minutes', () => {
-      const startTime = Date.now() - 61 * 1000 - 10 * MINUTE;
-
-      const result = getIsDelayed(StatusTypes.PENDING, {
-        startTime,
-        estimatedProcessingTimeInSeconds: 60,
-      } as BridgeHistoryItem);
-
-      expect(result).toBe(true);
-    });
-
-    it('returns false when current time exceeds estimated completion time, but less than 10 minutes have passed', () => {
-      const startTime = Date.now() - 61 * 1000;
-
-      const result = getIsDelayed(StatusTypes.PENDING, {
-        startTime,
-        estimatedProcessingTimeInSeconds: 60,
-      } as BridgeHistoryItem);
-
-      expect(result).toBe(false);
-    });
-  });
-
   describe('bridge snapshots', () => {
     it('should render completed bridge tx', () => {
       const { queryAllByTestId, getByText } = renderWithProvider(
